+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons 
 - [#534](https://github.com/Refhi/Weda-Helper/issues/534) - Ajout d'un bouton "+1clickVSM" dans la page des antécédents pour créer un VSM en un clic.
 - [#527](https://github.com/Refhi/Weda-Helper/issues/527) - Ajout d’un système de recherche rapide et floue (tolérant aux fautes de frappe) dans les motifs d’arrêt de travail AATI.
 - Les motifs d’arrêt de travail AATI sont désormais classés par ordre alphabétique.
+- [#535](https://github.com/Refhi/Weda-Helper/issues/535) - Ajout d’un message d’aide à la cotation pour GL1, GL2 et GL3 dans la page des FSE.
 
 ## Fix :
 - le bouton "AT sans CV" dans la page d'accueil du dossier patient fonctionne de nouveau correctement.